Make TestBroadcastReceiver inner class of ShadowApplicationTest.
