--- robolectric/src/test/java/org/robolectric/android/TestBroadcastReceiver.java
@@ -1,16 +1,0 @@
-package org.robolectric.android;
-
-import android.content.BroadcastReceiver;
-import android.content.Context;
-import android.content.Intent;
-
-public class TestBroadcastReceiver extends BroadcastReceiver {
-  public Context context;
-  public Intent intent;
-
-  @Override
-  public void onReceive(Context context, Intent intent) {
-    this.context = context;
-    this.intent = intent;
-  }
-}
